PY-64124 Match expected positional- and keyword-only parameters with regular parameters
The original false positive was caused by the protocol NestedSequence in matplotlib
defined as:

class NestedSequence(Protocol[_T_co]):
    def __getitem__(self, key: int, /) -> _T_co | NestedSequence[_T_co]: ...
    def __len__(self, /) -> int: ...

which was used to annotate parameters of matplotlib.pyplot.plot and was considered
incompatible with the builtin list.

Skipping positional- and keyword-only parameter separators of the expected callable
is a workaround until we have a comprehensive mechanism for matching signatures.

GitOrigin-RevId: 93d8bb4c6c4405d0e24b5f98152a461691f6197e