[wsm] Fix project reuse in light test
After the migration from MockSDK we faced the situation that light tests don't reuse instance of project if we use `SimpleLightProjectDescriptor` and its derivatives. The root cause is in SDK dispose thus we need to introduce the mechanize to compare old disposed SDK and the new one and if they are equal, replace it. To achieve it we need to save roots before dispose and reassign the SDK for the reused project descriptor.

GitOrigin-RevId: aeadd2c119f941650bb9b343b6bd83c6c812ad9f