IDEA-338714 Apply view settings in ScopeViewTreeModel constructor
These settings were previously only applied in invalidate(),
which caused the model to use the wrong settings initially.
Then they would be updated on any refresh, whether caused
by switching the scope pane, like in the ticket, or by some
other event such as a file status change. This led to the tree
changing its structure in an unexpected way.

Fix by applying the settings in the constructor.

GitOrigin-RevId: 7e313e9475b56da2d4f5a04fef861dab8ab22fcd