IDEA-333465 Disable autoselect on mouse move for Show Usages popup in screen reader mode
VoiceOver cursor was focusing on the whole ShowUsagesTable instead of the selected row, because ShowUsagesTableCellRenderer is based on SelectablePanel, which has default Panel accessible role (AxIgnore on macOS), and therefore cannot be focused by VO. If VO is configured with mouse pointer following VO cursor, mouse pointer will stay at the center of the table, and with enabled autoselect on mouse move for the table, it will always select the same row in the middle when trying to select a different row. Cell role cannot be changed to, e.g. Label, due to issues with VO cursor in tables (JBR-6344), so the autoselect on mouse move is disabled instead.

GitOrigin-RevId: 91b429990b190e8090cbaae1805132e5edc427d0