IDEA-334968 Disable selecting items in Git Branches popup by mouse move in screen reader mode
Due to issues with VO cursor in trees (JBR-6345, IDEA-338961), selecting items on mouse move causes VO to be stuck in a loop between items if VO is configured with a setting for the mouse pointer to follow VO cursor. As a quick fix, selecting on mouse move is disabled, since screen reader users mostly don't use mouse anyway.

GitOrigin-RevId: 7b834b4aa879380c6530e207c74f5c3433ccb495