IDEA-338980 Extend repaint area in TreeUtil.repaintBounds
The actual bounds on Windows is a flip of a coin: it could
be 1 px more or 1 px less. Even if we were to somehow
calculate them correctly here, there's no telling whether
it won't break someday when something is changed somewhere.

The most robust and safest fix is to simply repaint a bit more
than needed.

GitOrigin-RevId: 2730f50bcf8c7b07ceeba7aa3c395e5c08c10403