Introduce GeneralCommandLine.withEscapingForLocalRun for disabling command line mangling
Although GeneralCommandLine was designed a long time ago for running commands on the local machine, it's widely used for building and running commands on remote environments, namely WSL.

The host and the remote operating systems may vary. Command line escaping for running over cmd.exe produces corrupted arguments in terms of Posix.

The new flag allows to disable the arguments and is supposed to be used along with GeneralCommandLine.setProcessCreator.

GitOrigin-RevId: 0f90bb4191205ea3fcd5f1db31a8f02cfa679f73