IDEA-336199 IJent: implement uploading a binary via shell and use it in WSL
Initially, the IJent binary was launched in WSL through a Windows drive. A bug was found quickly: after the binary is recompiled, it segfaults in WSL. Most likely the problem was in WSL, because simple copying of the same binary to a new place resolved the problem. Also, restarting WSL helped.

The same situation with replaced binaries can appear in production after updating an IDE to a fresher version.

Instead of dealing with problems of WSL, it was decided to copy the binary into containers and run them directly.

The new bootstrapping shell script doesn't require anything except a single shell session, which decreases probability of other alleged file synchronization bugs. Moreover, such approach may be used later for running IJent over SSH.

GitOrigin-RevId: 69d1708babf151a6a640cdfb997711afbb0fe762