[wsm] Add method to calculate name of the SDK used in tests without instantiation
Due to the migration from MockSDK, we can't blindly create a lot of instances otherwise, it will lead to the instantiation of VirtualFilePointers, and thus, we need to dispose of them. In the old implementation, no one cares about it, but now we need to avoid the creation of redundant instances of SDK or dispose of them directly. In this case, I've introduced the method to calculate the SDK name without instantiation, unless it's needed. Without it, we'll get an exception about the undisposed root in   `KotlinSuspiciousPackagePrivateAccessInspectionTest28`

GitOrigin-RevId: 914bdc733a17768124f832bd60b82c98b4d61819