[terminal] IDEA-339096 Interrupt running command by Ctrl+C instead of selecting line at caret
It is specific of Windows and Linux, because Copy actions there has the same Ctrl+C shortcut. And we needed the unique action, that will handle all the cases when Terminal Copy actions are disabled. To not let EditorCopy action execute and select the line at caret.
Implemented TerminalInterruptCommandAction as a local action, because for now there is no need to show it in a Keymap.

GitOrigin-RevId: 08b1cb21b228c400ac5a9750b81a2a5a97989034