IDEA-339161 Consume close clicks in the file name widget
To avoid the popup showing up on a close click,
we need to consume it before the corresponding listener
is invoked. Since it's the processMouseEvent method
that's responsible for calling listeners, we override it
and make sure the event is consumed and the super
method is not called if it was a close click.

GitOrigin-RevId: 86e6fb8c0f757b3a3a393afbdfd3e64d4437d265