IDEA-339205 Set the size to preferred size after packing dialog wrapper
Due to the way Window.pack() is implemented and how components
with soft wraps calculate their height, there's a sort of chicken and egg
problem. Window.pack() first sets the dialog size to the preferred size,
but it doesn't account for soft wraps because the width isn't known
at the time the height is calculated. Then validation is performed,
and during it the actual preferred size becomes known. But it's
never actually used by pack().

Fix by overriding pack() and setting the size to the preferred size
after packing. If the preferred size doesn't change, it's effectively
a no-op, but if it does change, at least we account for this change.

To avoid calling our own setSize() overload, be careful to
call the super.setSize() overload that doesn't delegate to
one of our own.

GitOrigin-RevId: 442990ba5ab717918032d1721c78b2ebfa43c4fa