GTW-6993 Fix Related symbols popup is shown in the wrong position in RD
In RD `guessBestPopupLocation` returns `ClientRelativePoint` which is processed in a special way to ensure popup position is properly calculated.
Replacing this point with standard `RelativePoint` causes popup to be shown in the wrong position.

In `GotoRelatedSymbolAction` however there is no need to replace the `ClientRelativePoint` with `RelativePoint` as `popupLocation.getPoint().translate` alters the actual point (not its copy) therefore `new RelativePoint` may be safely removed

GitOrigin-RevId: 760f2d34459156755604b62b2ed9a4f0479e7d5f