IFT-525 Add Java or Kotlin onboarding tour promoter
Created the separate `intellij.kotlin.onboarding-promoter` module for the promoter extension.
This module is not bundled to Kotlin plugin because the promoter should be present only in IDEA Community/Ultimate (when Kotlin and Features Trainer plugins are enabled).
The name of this module is prefixed by `intellij`, because otherwise `KotlinModuleConsistencyTest` will fail. Also, there will be a lot of failed tests because this module won't be loaded in `ClassPathXmlPathResolver.resolveModuleFile` (`intellij` prefixed module descriptors can be missed when loading, by `kotlin` is not).
IJ-CR-121172

(cherry picked from commit 4b81ff2bfc0b60148e8b91fd35539edd113ab787)

GitOrigin-RevId: b48ef82fd9552c6e4a66d2dd6be0304c0747691c