[kotlin][scripting] KTIJ-27924 K2: MISSING_DEPENDENCY_CLASS on build.gradle.kts highlighting
Solution introduced to solve the KTIJ-27280 problem (8ab666a7) lacks an important point.
The thing is that `buildSrc` dependencies intersect with the ones from the `build.gradle.kts`.
During the analysis, PSI-element of a script binary dependency is "attached" to the `buildSrc`
module (this is how [1] function works). Given that, we calculate corresponding `KtModule`
incorrectly.

---------------------------------------------------------------------------------------------
[1]: ModuleUtilCore.findModuleForPsiElement()
[2]: ProjectStructureProviderIdeImplKt.calculateKtModule()

^KTIJ-27924 fixed

GitOrigin-RevId: 62d6f3953474b2acda684030f8995a780927718a