[wsm] Fix python tests
In general both of this two tests depends on the cache field by the execution of other tests e.g `PyUnresolvedReferencesInspectionTest.testMetaClassMembers` depends on the execution `PythonCompletionTest.testSmartFromUsedMethodsOfString`. `PythonCompletionTest` will fill PyPackageManagersImpl.myProvidedManagers`(internal cache) that cleans after the SDK dispose. Due to the migration from MockSDK to the real one this cache invalidates on each SDK recreation. But anyway TESTS SHOULD NOT depend on the order of tests

GitOrigin-RevId: 72f0a13a3a10019efcac37ca8de555a8a603b93a