IDEA-339309 Extract EditorColorsScheme.isDefaultForTheme extension
In order to avoid knowledge duplication, let's extract this part from
the implementation of the LaF manager to whatever place seems
convenient. Unfortunately, it's not possible to include this
in the EditorColorsScheme interface, as the necessary methods
aren't available in that module. So the counterpart, the theme,
seems to be the most logical place to put this.

This method is added now because it'll be needed to mark
the default scheme in places where the scheme can be selected,
as a hint of which of the schemes is associated with the current
theme by default.

GitOrigin-RevId: 09d63dce14b46361ff3783e28292fe25213e81d8