PY-63989 Introduce PyClassNameIndex.findByQualifiedName and PyFunctionNameIndex.findByQualifiedName
Replace PyClassNameIndex.findClass usages with this more descriptive API and remove the deprecated method.
Additionally, the new methods delegate to their caller control over handling several elements matching
a qualified name and the scope, in which the search is performed. If one wants to exclude python-skeletons,
they should explicitly request such a scope, e.g. with PySearchScopeBuilder.

This API is consistent with the one in PyModuleNameIndex.

GitOrigin-RevId: c128e6eceacbc39226504db018a780c09ed00adc