IDEA-299661 changed default prompt attirbute to the SYSTEM_OUTPUT
Not quite sure why it was set to the `USER_INPUT` when it is come from the process, so it should be a `NORMAL_OUTPUT` or a `SYSTEM_OUTPUT`. Feels more like `SYSTEM_OUTPUT` because it is just a markup, not the real information provided by the process.

Previous commit for this issue is intended to properly handle mix of the user input and process output and this unsent marker with `USER_INPUT` type caused sending it to the process.

GitOrigin-RevId: db98151476aadde72432fcaa74a353935e76b3b1