[groovy] properly reset classloader caches after stub generation is finished (IDEA-322782)
If class loaders are used to resolve references to classes in groovyc, it may happen that class file for groovy file is compiled from generated stub after groovy stub generation finished. To avoid "Unable to load class" error, we need to clear caches of the class loader which was created before stub generation started. Also, we need to disable the usage of classpath.index files which are updated only when the module chunk is fully compiled.

Second compilation is still needed in such cases, but given that it's performed automatically, it doesn't affect users much.

GitOrigin-RevId: 53b265c8966df06900e6b05558ce01ea8dfa77e9