IJPL-411: FileBasedIndexImpl.scheduleFileForIndexing should invalidate ALL indexes for file, not only REQUIRED according to indexing rules
This partially fixes IDEA-337316 (reduces a probability of an error), where a file was invalidated according to indexing rules in one project, and later it was indexed according to indexing rules of another project, and this indexing skipped some actually outdated indexes (e.g. Stub) because they were not invalidated properly.

GitOrigin-RevId: 506b9f94bdeadbc3c8c62e4f5ca496741b1245a0