[JavaScript Debugger] WEB-64044 Can't see inherited properties and debugger's "invoke getter" results in TypeError
Now, when loading object properties, 2 requests are sent:
- Runtime.getProperties(ownProperties = true, ...)
- Runtime.getProperties(ownProperties = false, accessorPropertiesOnly = true, ...)
This behavior aligns with Chrome DevTools and VS Code.

Also fixed getters evaluation in object prototypes. The problem was caused by an incorrect lookup for a non-prototype context: during the lookup, only the __proto__ field was checked, which was deprecated and replaced with [[Prototype]] internal property.


Merge-request: IJ-MR-121432
Merged-by: Shumaf Lovpache <Shumaf.Lovpache@jetbrains.com>

GitOrigin-RevId: eea9680cae3e994fe32354ea2d2f4194fcb5c51b