IDEA-337454 add ToggleCamelSnakeCase as known duplicate
Say we have for `ToggleCamelSnakeCase` action registration:

```      <keyboard-shortcut keymap="NetBeans 6.5" first-keystroke="alt shift K"/>
      <keyboard-shortcut keymap="$default" first-keystroke="alt shift U"/>```

Does it mean that the action will have two shortcuts if `NetBeans 6.5` is selected — a custom one (`alt shift K`) and from a parent `alt shift U`? Or only a custom one?

The surprising aspect here is that the order of lines matters. In the old implementation of `ActionManager`, we build a model as we read; in a new implementation, we separate the reading of the model and model application (as it should be).

So, in the old implementation, when we execute `boundShortcuts ?: parent?.getShortcutList(id)?.map { convertShortcut(it) } ?: emptyList()` and keyboard shortcut for `NetBeans 6.5` _before_ `$default`, we have here empty list. In the new implementation, we have a non-empty list (well, obviously, as a model for default keymap is applied first, as we create parent keymap first).

So, the question is — what's the correct behavior here? Given we have `replace-all` and `remove` flags (see [dev guide](https://plugins.jetbrains.com/docs/intellij/plugin-configuration-file.html#idea-plugin__actions__action__keyboard-shortcut) about it), looks like that a new behavior is a correct one (yes, we have tests for that, `IdeaKeymapsTest`, it includes Ruby plugin).

GitOrigin-RevId: 9b7cf0d37dcd2fdd0ac866297bfde10c3ad2c18f