IDEA-246330 Fix autoscroll a file located in a nested module
The cause of the bug: ModuleGroupNode only checked if the given file's
module is located in this group. However, the module may not belong
to this group, but its content root may be located under a content root
of another module which does belong to this group. In this case the file
will be shown under this group, indirectly, but the node will still return
false and the file will not be found.

The fix: not just check the modules the file belong to, but also check
the modules that contain the content roots of the said modules,
recursively. If any of such filesystem-wise parent modules are located
under this group, return true to ensure that the file is eventually found.

Implementation note: to prevent infinite recursion be very careful
not to check the modules that were already added to the set,
because all kinds of crazy situations are possible. For example,
a module may have a content root that's located under another
content root of the same module.

GitOrigin-RevId: f2f9bdb5baddf75d63eb0383f37ff481e1c59d47