IDEA-304699 New ScopeModel API
To fix slow ops, first create and implement a new API
using the old model implementation. Then we can switch
to a more efficient one later on with a registry key.

To keep compatibility with Search Everywhere which relies
on blocking calls for now, introduce a deprecated blocking method
for this new API. It's not supposed to be used anywhere else but
Search Everywhere.

The next step is to implement the same API using a new
coroutine-based model.

GitOrigin-RevId: 27daabd16378e6d867d7f9d7cbede07af19ca61a