IDEA-304699 Implement CoroutineScopeModel
The first step towards fixing slow ops related to scope comboboxes
and action groups. At this point most slow ops caused by
scope providers should be fixed, but there may be some unsafe UI
access now. This should be double-checked and tested properly.

There's a bunch of slow ops left in getPredefinedScopesAsync,
specifically in collectContext. These should be fixed separately,
probably a suspending equivalent of this API is required.

Because the code is now a mix of EDT and BGT code,
wrap the data context in an async data context.
It obviously still won't solve the issue of mixing UI
code with BGT code inside those scope providers,
if any, which should be fixed separately if anything pops up.

Add a registry key to turn on the new model, to be able to easily
turn it off in case anything goes wrong. This is still an experimental
implementation.

GitOrigin-RevId: 48c23e8423ee9b20ca57dcad5ea65571fbbe7d64