KTIJ-24647 Update KotlinSettingsEntity to include missing properties
The KotlinSettingsEntity entity has been extended to include three new properties - 'externalSystemRunTasks', 'version', and 'flushNeeded'. The 'externalSystemRunTasks' is needed for Gradle import of MPP projects. The 'version' is needed for opening old facets from disk. And the 'flushNeeded' property is needed for support of legacy implementations of read-and-write facet logic with realization changes on disk

GitOrigin-RevId: f10619fe167ad574681edf86c74ab3cb41bcd2f2