[android] IDEA-340049 Fix: Android Project Structure not loading
Add registryKey for gradle.build.toml support in android-plugin.xml to make Android plugin compatible
with IDEA 233.11799 release.

The registryKey was created in 233 branch after the IDEA 233.11799 release branch was created,
and because of this currently released plugin isn't compatible with IDEA 233.11799 release,
since it tries to use registry key which is not available in IDEA.

Since Android plugin is only component that uses the mentioned registerKey,
the registerKey is registered inside plugin.xml file instead of registry.properties file.

GitOrigin-RevId: a40d4ea4235385d065795e162f307593137e5cba