settings: don't warn about duplicate schemes: IDEA-321940
There can be different Themes with the same editor scheme inside, for example, ExperimentalLight and ExperimentalLightWithLightHeader, both of each reference the same editor scheme: /themes/expUI/expUI_lightScheme.xml with the schemeKey equal to Light.

A correct fix would be to not load the scheme twice, but preserve and reuse it. Since it is not that easy to fix it (an incorrect attempt was made in 301998ec and reverted in 02d2437b).

Until the proper fix is made, LOG.debug can be used to avoid a noisy warning on every start.

GitOrigin-RevId: 59766f514b51e9cc8a7436ee6787311984fdadeb