[wsm] Fix firing roots change event
Introduce a workaround for the cases where or `ProjectJdkTableImpl` doesn't use, so the entities create in the air, but any way we need to keep old contract and fire roots change event. Example of such case: `ServerProjectJdkTable` with enabled new implementation we have Sdk(with entity in it), but table stores SDKs in the list(not in the WSM).
Failed test: `com.jetbrains.python.PythonAnalysisToolSanityTest`

GitOrigin-RevId: 046545788c5d9feaa4d7885fd758515e329c05c1