IDEA-339828 IJent: dev build IDE downloads IJent from Maven repository
This commit removes a lot of doubtful code for borrowing IJent from a host IDE. Now, build scripts check if it is a TC agent or a developer machine, and the scripts download IJent on developer machines from a Maven repository.

There's no specific reason to use exactly Maven for delivering content. Any other resource server could be used, but some convenient utilities for using Maven as CDN already exist.

The version of the IJent is taken from build/dependencies.properties. A special job on CI updates that file asynchronously. Therefore, there's still a race between the moment of actual code update and the version update. However, similar races existed anymore, and the whole situation became significantly better. On the other side, occurence of such races would be quite painful for nightly installers on CI and absolutely unacceptable for EAP/Release builds. That's why this approach is used only for Dev Build.

GitOrigin-RevId: 8291c74afed61d330cfdfd7ede242e918a97dcc4