[workspace model] IDEA-332340 Explicitly set an SDK version at creation
The old and new behaviour of SDK is slightly different. The main difference in  version calculation(`LegacyProjectJdkDelegate.getVersionString`). In old implementation if version was not explicitly set it calculates automatically from the path when you are requesting it. But in the new implementation we don't have such implicit behaviour, so you need to specify it manually otherwise we will get an exception `Cannot Detect SDK Version`

GitOrigin-RevId: d85dde3c918de6ead861789f10444cc585455fef