[gtw] use the embedded client when connecting to a compatible host from Gateway started from IDE (RDCT-822)
If the Gateway process is started from an IDE, and it connects to a host with the same build number, we can use the client embedded in the IDE's installation instead of downloading a standalone one.

This is implemented by reusing existing ExtractedJetBrainsClientData class for such cases to avoid changes in the API. To implement this properly, ExtractedJetBrainsClientData should be refactored to support the cases with downloaded client and embedded client explicitly, and its instance should be passed from functions which download the client to functions which run the client.

(cherry picked from commit 18a4617d38c1a54d2c8e288e3122257a8789d01a, IJ-CR-122194)

GitOrigin-RevId: 1d7602cde07aec62125a6eaacc8128ad489ef3c1