Fix false positive failure in `IjentWslTest.WslIjentManager must restart IJent if it exits`
There's a coroutine that waits for the exit code of IJent and throws an error if the exit code is not zero. It's a useful check, but this check is undesired in the test that deliberately kills the IJent process for checking how the system reacts on that.

Two changes help to control the problem:
* Instead of passing exceptions to the default error handler, `WslIjentManager` explicitly logs errors.
* The logged errors are handled in the test using `LoggedErrorProcessor`.

GitOrigin-RevId: b374b7dbbb3b12d48b25d6b91ac9648d098bc8a8