IDEA-323453 Fix VoiceOver not reading complete lines in the editor after switching tabs
Disabled ACCESSIBLE_SELECTION_PROPERTY changed event for JBTabsImpl on Mac, because with this event, VoiceOver focus moves to the tab itself, and as a result it doesn't work well with the tab content (Editor text in particular). This is a workaround, because the event might be useful, but the actual root cause is unclear.

GitOrigin-RevId: d1f557954c749465d1c9456e8d1affaf2fad068a