K2: implement create-Kotlin-callable-from-usage-in-Kotlin
We have a powerful architecture that supports cross-language for
create-from-usage like creating Java elements from usage in Java and
Groovy. They share the same CreateMethodAction based on
CreateMethodRequest. When we want to add a create-Java-element
quickfix from a language, we can simply implement a request class that
is a child of CreateMethodRequest.

Likewise, we already have a skeleton classes to support the same
cross-language feature for create-Kotlin-element from usage, which is
CreateKotlinCallableAction. It creates Kotlin callable based on a given
request.

This commit implements K2 create-from-usage (previously, we had
CreateCallableFromCallActionFactory and CreateCallableFromUsageFix)
based on CreateKotlinCallableAction and request. Currently,

- Correct action messages for quick-fixes -> working
- [WIP] Function creation
  - Container selection to inserts created functions -> working
  - Capturing parameter types information -> working
  - Parameter name selection -> working
  - Capturing return type information -> working
  - [WIP] Rendering types
  - [TODO] Template
  - [TODO] Shortening reference
  - [TODO] Handling type parameter
  - [TODO] Moving caret after inserting created function