K2 create-from-usage: Support template for parameters and return type
Based on the template, this commit support the selection of parameter
names, parameter types, and return type. Now K2 create-from-usage for
functions has the same behaviors as K1 for "abstract" call test cases.
It was also tested in an instance of IJ as well.

Now the minimal K2 create-function-from-usage works:
- Correct action messages for quick-fixes -> working
- [WIP] Function creation
  - Container selection to inserts created functions -> working
  - Capturing parameter types information -> working
  - Parameter name selection -> working
  - Capturing return type information -> working
  - Rendering types -> working
  - Template -> working
  - Shortening reference -> working
  - Moving caret after inserting created function -> working
  - [TODO] Handling type parameter