IDEA-324815 Isolate expand/collapse code in Tree
To fix the performance issues with bulk expand/collapse
operations, the first step is to make sure that all code
related to these operations is located in our codebase.
To achieve that, override all JTree methods that deal
with expand/collapse states and implement them in our Tree.

Create a registry key to enable/disable this new API
at a whim, so we can always fall back to the default
implementation as if nothing was ever changed.

GitOrigin-RevId: 37fe8bf332ba09f835a6fe77395e005ee55dadb7