GTW-7039 BackendRemoteSettingsHost. on EDT at startup causes 5 seconds UI freeze
Get EditorColorsManager.getInstance().getGlobalScheme() before enumerating all schemes because in the case of RD the code in UnattendedHostRemoteControlHandler can change the current scheme to a temporary one after we collected all schemes into the list mySchemes. It led to the assert written in the end of initAll(). Actually it's a workaround, this code should be made thread safe somehow

GitOrigin-RevId: 3913bb60f8bdb41fdcfdaf9fd82d6c50b6bdef34