set "performGroup" flag before performing action-groups
In backend performing groups we avoid two server calls, hence empty `beforeActionPerformedUpdate`. If the "performGroup" flag is set only in `update` then we lose it. Fortunately, in menus and action popups, perform is only called for performing groups, and we can set the flag without additional checks.

Note that the original presentation from `update` is not preserved to avoid memory leaks.

Fixes empty popup on choosing Debug in the run line marker menu in a rem-dev scenario.

GitOrigin-RevId: 1f064b5d1326c6322a5187ba539ce865fcfb10e8