IDEA-330628 Add support for .compact values in themes
Because the key is used to detect the value type,
we need to remove the .compact suffix for detection
to work properly now. Note that numeric types
are supported automatically anyway because their
type is detected on the JSON level, so no guesswork
based on the key name is needed there.

GitOrigin-RevId: 93e7271d3c26c5fe43db5410893339aebf0b378b