K2 create-from-usage: Avoid unnecessary KT to LC/PsiType conversion
The creation of Kotlin functions from Kotlin usage does not need LC and
PsiType because both sides use KtClass and KtType. To avoid the
unnecessary conversion between KtClass and LC and conversion between
KtType and PsiType, this commits adds wrapper classes for KtClass and
KtType that are child classes of JvmClass and JvmType. See
JvmClassWrapperForKtClass and JvmTypeWrapperForKtType.

GitOrigin-RevId: 13b67b3573d8acef98d41aa9681cfb3070767893