[MERMAID-84][MERMAID-86] Explicitly set page dimensions to the default size used by Puppeteer in the mermaid-cli
It turns out that some diagrams (namely pie chart) have their dimensions hardcoded. It seems that such hardcoded dimensions were only tested with the default viewport dimensions provided by the Puppeteer (800x600). So, to fix most of the issues with positioning, we have to explicitly set the page dimensions to match the defaults from the Puppeteer.

GitOrigin-RevId: ab8788e7177e98d304eb04df4b9c2ed976fe0cc9