[RDCT] GTW-6734: Do not replace data context in Evaluate Expression handler for remote clients
This data context guess and replacement is a hack that allows Evaluate Expression action to work properly in the standalone IDEs. But this doesn't work in CWM and Remote Dev if this action is called from a remote client. Data context in such cases are properly formed by the RDCT Core code, and it should not be replaced by a focus-based guess.

GitOrigin-RevId: f19cca7709c3f4b3009daaabe335560729147703