[kotlin] Remove intersection with project test scope when creating scope for KtSourceModuleByModuleInfo
This intersection is not required as `ideaModule.moduleTestSourceScope` is expected to be a subset of `GlobalSearchScopes.projectTestScope(project)`. At the same time, this intersection is the cause of a performance degradation.
With abe9024c90c return value of `extractFileEnumeration` for `IntersectionScope` created in `KtSourceModuleByModuleInfo` becomes `null`, and degradation in completion performance appears, because optimization previously provided by VirtualFileEnumeration no longer works.
^KTIJ-28083 Fixed

(cherry picked from commit 7365a5b7d92ad279cebfef788e9e830732848791)

IJ-CR-122670

GitOrigin-RevId: c1175a6ee90b6c5413b36df9bdbaba62439abb73