[kotlin][scripting] KTIJ-16925: resolve scope provider affects libraries analysis
`ScriptDependenciesResolveScopeProvider` is aimed for script dependencies only.
In cases when library binary/source file was also among script dependencies, it
was resolved in the wrong scope - script's one.

With this commit, we сut such cases out, checking the file against library
roots first.

Due to KT-61944 we have to make a choice. In case a dependency is available
both in project libraries and in script dependencies, we don't know what scope
to prefer (library module vs script dependencies).

Now, we "say" that we prefer correct analysis of regular source files to scripts.

^KTIJ-16925 fixed

GitOrigin-RevId: af7b30c74c75d606a9884632181107067a667081