[terminal] IDEA-339408 Show documentation for commands
It was required to edit the way of how command specs are loaded and stored to access the command descriptions without loading the whole command specification. So, in the new implementation, we first load short command specifications (names, description and loadSpec reference) for all commands. And these descriptions are used for showing the documentation.

GitOrigin-RevId: ae46abcd1f1a8f1f40f4a43e56c8d824c247df04