[split chooser] fix bug when the action closes the tab instead of moving the tab #IDEA-321319 fixed
Previously a `null` argument was passed to `window` parameter of `FileEditorManagerImpl.openFile`. So `openFile` tried to guess in which window it should open the file.
It guessed the initialWindow instead of activeWindow, because the file editor is still open there.
However, several lines below we close the editor `initialWindow.closeFile(file = file)`, so effectively we just close the editor.

GitOrigin-RevId: f302f89b9bfc2b4f1d876bbedc01c0c2d6a7bfd9