[python] Fix env tests for pycodestyle.py
It turned out that env tests don't configure the discovered temporary SDK as the module
SDK automatically. As a result, we considered files being checked written in Python 3.13
and launched the latest version of pycodestyle.py on actual Python 3.7.

Now, I set the module SDK explicitly in PyPep8Test and additionally use the version
of the worker interpreter for running pycodestyle.py instead of the version of the
file being analyzed in Pep8ExternalAnnotator, because these two might not match.

GitOrigin-RevId: 99a12a66e5a1a0360189d8ec04b339b59af93e87