[performance-tests] wait for the full log before computing file history
If log is not initialized before file history is requested, the initialization process is started in parallel. The duration of computing file history is dependent on the moment when initialization process is finished, since without the full data pack building the visible graph is skipped.

Waiting for the full log allows to get results which are more close to the real usages and add stability to the test.

GitOrigin-RevId: 73059e27bcee4dbfee42e9cd4213a153d169836a