[core, kotlin] disable plugins that are incompatible with the Kotlin Plugin in K2 mode
The solution is temporary until a better one is found, which does not require registering a fake `supportsKotlinK2Mode` extension and modifying the plugin loading code.

The following code can be added to `plugin.xml` to allow it to be used in the K2 Mode. This is not a public way of doing this, and it will probably be removed for the 2024.2 release.

```
<extensions defaultExtensionNs="org.jetbrains.kotlin">
  <supportsKotlinK2Mode/>
</extensions>
```

^KTIJ-24797 fixed

GitOrigin-RevId: 8ea078ee7d567c5c31fa7011bcf817628a78a6a7