GTW-6938 Hack for 'Open project in this window' leads to Welcome screen
OpenFileAction opens a project under cancellable progress dialog (`runWithModalProgressBlocking`).
If the dialog gets closed for whatever reason, it is considered as though the process has been canceled and project opening should be interrupted.

The dialog is sent to frontend using LUX and shown there with `IdeFrame` as its owner, so the dialog will be closed when `IdeFrame` is disposed.

`IdeFrame` is disposing when the last project is closed, which happens when a new project is opening in the same window (firstly, an old project is closed, then new is opened).

Standalone IDE prevents this from happening by reusing the same IdeFrame (`withFrameReuseEnabled`)
Frontend, however, cannot use this flag, as it is unaware of a new project will be opened soon after.

Due to the fact that this has to be fixed in 233 and there is no solution which wouldn't require major changes probably both in RD and platform, we've come up with a hack safe enough to be used for now.

The hack is that we don't allow project opening to be a cancelable process (coroutine). If it's not cancelable, then closing the dialog won't prevent a project from opening.
The IdeFrame on frontend will be closed anyway and reopened with a new project after a few seconds.

GitOrigin-RevId: 9f3468fcf93b70db5136951ceebf2d23494ad1a5