update AssertJ
This commit updates the versions of AssertJ and Byte Buddy used in the project. It changes AssertJ core from version 3.24.2 to 3.25.1 and Byte Buddy from 1.12.21 to 1.14.11 across different components of the application to keep up with the latest updates. These updates may include bug fixes, new features or performance enhancements, thereby improving the overall code quality and efficiency.

GitOrigin-RevId: c9b4a77b2f9dc1ade38d77987273b6365c27a6b2