IDEA-342398 Fix modality in ImageFileService
When invoked from a modal dialog, the previous implementation
would never be invoked until the dialog is closed. If the image editor
is in the dialog, it would never show an image.

This service was initially supposed to work in any modal context.
It's because it used to be completely synchronous code before,
so it's really supposed to be invoked as soon as possible,
just without blocking the EDT anymore.

GitOrigin-RevId: 778f870a70a6078fdab893b91e1ea8ab42265ee2