project-wizard: don't suggest creating a Git repository when creating a module: IDEA-342473
Creating a Git repository is useful when creating a new project, but not when creating a module under an existing project. Creating a sub-repository for a module is a rare and advanced task and should be done explicitly and manually.

(cherry picked from commit 33ffb1b245fb804403e298e4a1d7399f3a9854e9)
IJ-CR-123430

GitOrigin-RevId: f79db697b610b18e3533af90b2c7141efacdd1d1