IDEA-341232 Fix the theme default hint for legacy themes
These themes don't specify a scheme ID, so the default schemes
are used: Darcula for dark themes, Default for light themes.
This is determined by defaultNonLaFSchemeName(), but we
can't use it directly because it's from the platform-impl module,
not available in the platform-api module. So just copy-paste the
logic there, which is, of course, far from perfect, but shouldn't hurt
that much, because it's a private implementation detail anyway,
and in the worst case (if that logic ever changes), nothing
terrible will happen: possibly the hint will just be wrong.
And since it's legacy stuff, any changes are unlikely anyway.

GitOrigin-RevId: aa4791e72c39cfbc08fe5f1cde09b51e1571bb26