IDEA-341521 Fix ToolbarDecorator preferred size for scrollables
ToolbarDecorator overrides JBScrollPane's getPreferredSize()
for its own reasons, but only if it isn't set explicitly.
However, it may also be set implicitly, through overriding
getPreferredScrollableViewportSize(). In this case we don't
want ToolbarDecorator to override it either, because it's
essentially the same as setting the preferred size explicitly:
the client code tells us what size the viewport should preferably
have, and we shouldn't get in the way.

Fix by checking if getPreferredScrollableViewportSize()
returns null and only overriding the preferred size if it does.

GitOrigin-RevId: 73c45e7e4eabad3a3f4679f2472946c7e6b971e6