GTW-6938 Allow RD to reuse frame when a project is opening in the same window
When a project is opening in the same window, standalone IDE reuses frame from the previous project instead of opening a new one.
It prevents window blinking and also some logic relies on the frame being reused.
More info about this logic can be found in the issue or the commit 9f3468fcf93b70db5136951ceebf2d23494ad1a5

In RD, however, when the project is closing, we don't know whether a new project is going to be opened and therefore the frame should be reused.

To support frame re-usages, I add new options to `close` method for a `Solution` model. These options can contain all necessary parameters/flags that can modify a project closing scenario.

To be able to retrieve the options, a solution is now closes only when `close` is fired instead of on lifetime termination event.

(cherry picked from commit b468aaa4de2b04d8aa3c9fd644101ef7780129b5)

GitOrigin-RevId: 379c20613957ce12db5e4efdbeaea2e6f1ab193e