IDEA-324815 Implement bulk expand/collapse API
There are three levels of implementation.

First, the new Tree methods, expandPaths and collapsePaths.
These update the state of the tree itself.

Second, when the state of every node is updated,
the tree fires an event that reaches its UI. This makes it
update the visible structure (the tree layout cache)
every time, even if we implement it as a bulk operation
on the tree side. To avoid this, create a new interface
for tree UIs that actually support bulk operations. Now the tree
can tell its UI delegate that there's a bulk operation in progress,
so it can postpone actual updates. Implement this interface
in DefaultTreeUI.

Third, the tree state itself, DefaultLayoutCache. To avoid
rebuilding it over and over again, introduce a new method
there that simply rebuilds the rows, loading nodes as needed,
but without screwing up state unrelated to expanded/collapsed
stuff.

In this new implementation, once a bulk operation is over,
the entire layout cache is updated in one go.

GitOrigin-RevId: 9098f32eaeff6c53f96f59353ba4def9c1b40b64