(IDEA-342602) Rewrite semantic search settings, change persistence approach, move to search-everywhere-ml plugin package
I started saving to disk only the flag that a user manually changed a setting value and the value itself in this case. The check of the experiment group now happens at runtime when we read the settings, so we always use only the relevant experiment plan.

I also moved search everywhere semantic settings to search-everywhere-ml plugin package, because it's more related to search everywhere than the general embeddings API that can be used across IntelliJ

GitOrigin-RevId: 409d06bfc103887fe04bb0ebdac10e32be7284e2